refactor: rename MobileFooter
